+++ tests/source/features detection/Test_HarrisCorner.cpp
@@ -17,8 +17,8 @@ TEST(Test_shakhbat_cv, HarrisCorner)
 
 	auto gray = qlm::ColorConvert< qlm::ImageFormat::RGB, uint8_t, qlm::ImageFormat::GRAY, uint8_t>(in);
 
-	size_t filter_size = 3;
-	size_t block_size = 3;
+	int filter_size = 3;
+	int block_size = 3;
 	float k = 0.22;
 	float threshold = 0.0000001;
